Added everything into One file
Everything is in Assignment1.py. Everything else is in extras. There is
still a problem with creating a path.
It seems that Dijkstras does not add a lot of runners to the heap. This
could be a problem with Dijkstas, the creation of the graph, finding
the proper vertices, or reading the path back. Still need to alter some
functions to match the declaration given on eclass.